Split up esm and cjs to pass down the proper require
`getMockMiddleware` needs to use a `require` function to refresh the
cache when loading a new `commonjs` file.

* If this package is using the "commonjs" version, it will use the
  default node require.
* If this package is using the "module" version, it uses the
  `createRequire` helper to load commonjs files, and allows clearing the
  `require` cache.

Because typescript cannot compile commonjs files that use `import.meta`,
the code is split up into two index files, which pass down the needed
`require` function.

To determine what file it will load, it will first check the extension
(either mjs or cjs), and otherwise will fall back to checking the
closest package.json file to see what the "type" is set to.